+++ src/WebJobs.Script.Grpc/MultiLanguageFunctionDescriptorProvider.cs
@@ -26,10 +26,10 @@ namespace Microsoft.Azure.WebJobs.Script.Description
         /// <param name="bindingProviders">List of <see cref="IScriptBindingProvider"/> instances.</param>
         /// <param name="dispatcher"><see cref="IFunctionInvocationDispatcher"/> instance.</param>
         /// <param name="loggerFactory"><see cref="ILoggerFactory"/> instance.</param>
-        /// <param name="applicationLifetime"><see cref="IHostApplicationLifetime"/> instance.</param>
+        /// <param name="applicationLifetime"><see cref="IScriptApplicationLifetime"/> instance.</param>
         /// <param name="workerInitializationTimeout">Worker initialization timeout.</param>
         public MultiLanguageFunctionDescriptorProvider(ScriptHost host, IList<RpcWorkerConfig> workerConfig, ScriptJobHostOptions config, ICollection<IScriptBindingProvider> bindingProviders,
-            IFunctionInvocationDispatcher dispatcher, ILoggerFactory loggerFactory, IHostApplicationLifetime applicationLifetime, TimeSpan workerInitializationTimeout)
+            IFunctionInvocationDispatcher dispatcher, ILoggerFactory loggerFactory, IScriptApplicationLifetime applicationLifetime, TimeSpan workerInitializationTimeout)
             : base(host, config, bindingProviders, dispatcher, loggerFactory, applicationLifetime, workerInitializationTimeout)
         {
             _workerConfig = workerConfig ?? throw new ArgumentNullException(nameof(workerConfig));